fix: require int argument for Conv2d

+++ models/resnet.py
@@ -78,7 +78,7 @@ class Bottleneck(nn.Module):
 class ResNet(nn.Module):
     def __init__(self, block, num_blocks, num_classes=10, widen_factor=1):
         super(ResNet, self).__init__()
-        self.in_planes = 64 * widen_factor
+        self.in_planes = int(64 * widen_factor)
         
         """
         # ImageNet downscaling
@@ -86,13 +86,13 @@ class ResNet(nn.Module):
                                stride=2, padding=3, bias=False)
         self.bn1 = nn.BatchNorm2d(64)
         """
-        self.conv1 = nn.Conv2d(3, 64 * widen_factor, kernel_size=3, stride=1, padding=1, bias=False)
-        self.bn1 = nn.BatchNorm2d(64 * widen_factor)
-        self.layer1 = self._make_layer(block, 64 * widen_factor, num_blocks[0], stride=1)
-        self.layer2 = self._make_layer(block, 128 * widen_factor, num_blocks[1], stride=2)
-        self.layer3 = self._make_layer(block, 256 * widen_factor, num_blocks[2], stride=2)
-        self.layer4 = self._make_layer(block, 512 * widen_factor, num_blocks[3], stride=2)
-        self.linear = nn.Linear(512*block.expansion*widen_factor, num_classes)
+        self.conv1 = nn.Conv2d(3, int(64 * widen_factor), kernel_size=3, stride=1, padding=1, bias=False)
+        self.bn1 = nn.BatchNorm2d(int(64 * widen_factor))
+        self.layer1 = self._make_layer(block, int(64 * widen_factor), num_blocks[0], stride=1)
+        self.layer2 = self._make_layer(block, int(128 * widen_factor), num_blocks[1], stride=2)
+        self.layer3 = self._make_layer(block, int(256 * widen_factor), num_blocks[2], stride=2)
+        self.layer4 = self._make_layer(block, int(512 * widen_factor), num_blocks[3], stride=2)
+        self.linear = nn.Linear(int(512*block.expansion*widen_factor), num_classes)
     
         self.activate = F.relu
         # self.activate = nn.LeakyReLU(negative_slope=0.1, inplace=True)
